Share the isolating class loader test fixture between core and generator
Both class loader isolation tests carried an identical child loader. Move
it to the core test jar, which the generator module already consumes
through the tests classifier.
